Move MC card emulation related stuff into separate folder
